Fixed pie chart
For a while, there's been an error with the pie chart - it displays the wrong stats.

I printed out the result variables and I'm pretty sure it's because the code for checking the winner was still being accessed if you click the 'play again' button.

I fixed this in the event handling method's if statement and reduced the branches - instead of checking for each individual RPS button, I can cast the event as a button and get it's text.

I've also made small changes throughout the codebase so maybe it will be easier to read and maintain in the future.